requirements.txt
streamlit>=1.20
pdfplumber>=0.7.5
requests>=2.28
dicttoxml>=1.7.4
python-dotenv>=1.0.0